src/requirements.txt
# # sudo apt-get install python3 python3-pip
# # virtualenv
# # python3-setuptools
# # python3-pip
# # python3-dev 
# # python3-venv 
# # git 

# Django>=2.2,<3.0
# djangorestframework
# markdown
# django-filter
# psycopg2==2.7.3.2
# # Pillow
# gunicorn
# # django-heroku
# django-tinymce4-lite
# django-filebrowser-no-grappelli
# django-crispy-forms
# django-cors-headers
# # in frontend folder:
# # npx create-react-app gui
# # npm install axios --save
# # npm install --save react-router-dom

astroid==2.2.5
autopep8==1.4.4
colorama==0.4.1
Django==2.2.5
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-filebrowser-no-grappelli==3.7.8
django-filter==2.2.0
django-tinymce4==3.0.2
django-tinymce4-lite==1.7.5
djangorestframework==3.10.3
gunicorn==19.9.0
redis==3.3.8
django-heroku==0.3.1
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
django-celery-results==1.1.2
# Pillow==6.1.0
django-suit==0.2.28
psycopg2==2.8.3
pycodestyle==2.5.0
requests==2.22.0
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
django-environ==0.4.5